Restrict number of inputs, must fit into uint64_t bitmask

+++ src/drivers/input/input_driver.hpp
@@ -19,7 +19,7 @@ struct Input {
   };
 
   // Configuration
-  size_t idx;
+  uint8_t idx;
   etl::string<20> name;
   bool invert = false;
   EmergencyMode emergency_mode;
